+++ src/mapleall/maple_me/src/me_value_range_prop.cpp
@@ -3230,17 +3230,17 @@ void ValueRangePropagation::AnalysisUnreachableBBOrEdge(BB &bb, BB &unreachableB
   }
   Insert2UnreachableBBs(unreachableBB);
   // update frequency before cfg changed
-  uint64 removedFreq = 0;
+  int64_t removedFreq = 0;
   if (func.GetCfg()->UpdateCFGFreq()) {
     int idx = bb.GetSuccIndex(unreachableBB);
-    removedFreq = bb.GetSuccFreq()[static_cast<uint32>(idx)];
+    removedFreq = static_cast<int64_t>(bb.GetSuccFreq()[static_cast<uint32>(idx)]);
   }
   bb.RemoveSucc(unreachableBB);
   bb.RemoveMeStmt(bb.GetLastMe());
   bb.SetKind(kBBFallthru);
   if (func.GetCfg()->UpdateCFGFreq()) {
     bb.SetSuccFreq(0, bb.GetFrequency());
-    succBB.SetFrequency(succBB.GetFrequency() + removedFreq);
+    succBB.SetFrequency(static_cast<uint32>(succBB.GetFrequency() + removedFreq));
     unreachableBB.SetFrequency(0);
   }
   auto *loop = loops->GetBBLoopParent(bb.GetBBId());
@@ -3623,9 +3623,9 @@ void ValueRangePropagation::RemoveUnreachableBB(
       condGotoBB.SetKind(kBBFallthru);
       // update frequency before cfg changed
       if (func.GetCfg()->UpdateCFGFreq()) {
-        uint64 removedFreq = condGotoBB.GetSuccFreq()[1];
+        int64_t removedFreq = static_cast<int64_t>(condGotoBB.GetSuccFreq()[1]);
         condGotoBB.SetSuccFreq(0, condGotoBB.GetFrequency());
-        succ0->SetFrequency(succ0->GetFrequency() + removedFreq);
+        succ0->SetFrequency(static_cast<uint32>(succ0->GetFrequency() + removedFreq));
       }
       condGotoBB.RemoveSucc(*succ1);
       DeleteThePhiNodeWhichOnlyHasOneOpnd(*succ1, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
@@ -3639,15 +3639,15 @@ void ValueRangePropagation::RemoveUnreachableBB(
         UpdateProfile(*condGotoBB.GetPred(0), condGotoBB, trueBranch);
       }
       condGotoBB.SetKind(kBBFallthru);
-      uint64 removedFreq = 0;
+      int64_t removedFreq = 0;
       // update frequency before cfg changed
       if (func.GetCfg()->UpdateCFGFreq()) {
-        removedFreq = condGotoBB.GetSuccFreq()[0];
+        removedFreq = static_cast<int64_t>(condGotoBB.GetSuccFreq()[0]);
       }
       condGotoBB.RemoveSucc(*succ0);
       if (func.GetCfg()->UpdateCFGFreq()) {
         condGotoBB.SetSuccFreq(0, condGotoBB.GetFrequency());
-        succ1->SetFrequency(succ1->GetFrequency() + removedFreq);
+        succ1->SetFrequency(static_cast<uint32>(succ1->GetFrequency() + removedFreq));
       }
       DeleteThePhiNodeWhichOnlyHasOneOpnd(*succ0, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
       condGotoBB.RemoveMeStmt(condGotoBB.GetLastMe());
@@ -3806,20 +3806,20 @@ bool ValueRangePropagation::ChangeTheSuccOfPred2TrueBranch(
   if (exitCopyFallthru != nullptr) {
     PrepareForSSAUpdateWhenPredBBIsRemoved(pred, bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
     size_t index = FindBBInSuccs(pred, bb);
-    uint64 edgeFreq = 0;
+    int64_t edgeFreq = 0;
     if (func.GetCfg()->UpdateCFGFreq()) {
-      edgeFreq = pred.GetSuccFreq()[index];
+      edgeFreq = static_cast<int64_t>(pred.GetSuccFreq()[index]);
     }
     pred.RemoveSucc(bb);
     DeleteThePhiNodeWhichOnlyHasOneOpnd(bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
     pred.AddSucc(*exitCopyFallthru, index);
     CreateLabelForTargetBB(pred, *exitCopyFallthru);
     if (func.GetCfg()->UpdateCFGFreq()) {
-      exitCopyFallthru->SetFrequency(exitCopyFallthru->GetFrequency() + edgeFreq);
-      pred.AddSuccFreq(edgeFreq, index);
+      exitCopyFallthru->SetFrequency(static_cast<uint32>(exitCopyFallthru->GetFrequency() + edgeFreq));
+      pred.AddSuccFreq(static_cast<uint64>(edgeFreq), index);
       // update bb frequency
       ASSERT(bb.GetFrequency() >= edgeFreq, "sanity check");
-      bb.SetFrequency(bb.GetFrequency() - edgeFreq);
+      bb.SetFrequency(static_cast<uint32>(bb.GetFrequency() - edgeFreq));
       bb.UpdateEdgeFreqs();
     }
     return true;
@@ -3855,19 +3855,19 @@ bool ValueRangePropagation::ChangeTheSuccOfPred2TrueBranch(
   mergeAllFallthruBBs->AddMeStmtLast(gotoMeStmt);
   PrepareForSSAUpdateWhenPredBBIsRemoved(pred, bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
   size_t index = FindBBInSuccs(pred, bb);
-  uint64 edgeFreq = 0;
+  int64_t edgeFreq = 0;
   if (func.GetCfg()->UpdateCFGFreq()) {
-    edgeFreq = pred.GetSuccFreq()[index];
+    edgeFreq = static_cast<int64_t>(pred.GetSuccFreq()[index]);
   }
   pred.RemoveSucc(bb);
   pred.AddSucc(*mergeAllFallthruBBs, index);
   if (func.GetCfg()->UpdateCFGFreq()) {
-    mergeAllFallthruBBs->SetFrequency(edgeFreq);
-    mergeAllFallthruBBs->PushBackSuccFreq(edgeFreq);
-    pred.AddSuccFreq(edgeFreq, index);
+    mergeAllFallthruBBs->SetFrequency(static_cast<uint32>(edgeFreq));
+    mergeAllFallthruBBs->PushBackSuccFreq(static_cast<uint64>(edgeFreq));
+    pred.AddSuccFreq(static_cast<uint64>(edgeFreq), index);
     // update bb frequency
     ASSERT(bb.GetFrequency() >= edgeFreq, "sanity check");
-    bb.SetFrequency(bb.GetFrequency() - edgeFreq);
+    bb.SetFrequency(static_cast<uint32>(bb.GetFrequency() - edgeFreq));
     bb.UpdateEdgeFreqs();
   }
   mergeAllFallthruBBs->AddSucc(trueBranch);
@@ -3938,9 +3938,9 @@ bool ValueRangePropagation::RemoveTheEdgeOfPredBB(
   if (OnlyHaveCondGotoStmt(bb)) {
     PrepareForSSAUpdateWhenPredBBIsRemoved(pred, bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
     size_t index = FindBBInSuccs(pred, bb);
-    uint64 edgeFreq = 0;
+    int64_t edgeFreq = 0;
     if (func.GetCfg()->UpdateCFGFreq()) {
-      edgeFreq = pred.GetSuccFreq()[index];
+      edgeFreq = static_cast<int64_t>(pred.GetSuccFreq()[index]);
       ASSERT(bb.GetFrequency() >= edgeFreq, "sanity check");
     }
     pred.RemoveSucc(bb);
@@ -3948,22 +3948,22 @@ bool ValueRangePropagation::RemoveTheEdgeOfPredBB(
     pred.AddSucc(trueBranch, index);
     CreateLabelForTargetBB(pred, trueBranch);
     if (func.GetCfg()->UpdateCFGFreq()) {
-      bb.SetFrequency(bb.GetFrequency() - edgeFreq);
+      bb.SetFrequency(static_cast<uint32>(bb.GetFrequency() - edgeFreq));
       size_t trueBranchIdx = static_cast<uint32>(bb.GetSuccIndex(trueBranch));
-      int64_t updatedtrueFreq = static_cast<int64_t>(bb.GetSuccFreq()[trueBranchIdx] - edgeFreq);
+      int64_t updatedtrueFreq = static_cast<int64_t>(bb.GetSuccFreq()[trueBranchIdx] - static_cast<uint64>(edgeFreq));
       // transform may not be consistent with frequency value
       updatedtrueFreq = updatedtrueFreq > 0 ? updatedtrueFreq : 0;
       bb.SetSuccFreq(static_cast<int>(trueBranchIdx), static_cast<uint64>(updatedtrueFreq));
-      pred.AddSuccFreq(edgeFreq, index);
+      pred.AddSuccFreq(static_cast<uint64>(edgeFreq), index);
     }
   } else {
     auto *exitCopyFallthru = GetNewCopyFallthruBB(trueBranch, bb);
     if (exitCopyFallthru != nullptr) {
       PrepareForSSAUpdateWhenPredBBIsRemoved(pred, bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
       size_t index = FindBBInSuccs(pred, bb);
-      uint64 edgeFreq = 0;
+      int64_t edgeFreq = 0;
       if (func.GetCfg()->UpdateCFGFreq()) {
-        edgeFreq = pred.GetSuccFreq()[index];
+        edgeFreq = static_cast<int64_t>(pred.GetSuccFreq()[index]);
         ASSERT(bb.GetFrequency() >= edgeFreq, "sanity check");
       }
       pred.RemoveSucc(bb);
@@ -3971,14 +3971,15 @@ bool ValueRangePropagation::RemoveTheEdgeOfPredBB(
       pred.AddSucc(*exitCopyFallthru, index);
       CreateLabelForTargetBB(pred, *exitCopyFallthru);
       if (func.GetCfg()->UpdateCFGFreq()) {
-        bb.SetFrequency(bb.GetFrequency() - edgeFreq);
-        exitCopyFallthru->SetFrequency(edgeFreq);
-        exitCopyFallthru->PushBackSuccFreq(edgeFreq);
+        bb.SetFrequency(static_cast<uint32>(bb.GetFrequency() - edgeFreq));
+        exitCopyFallthru->SetFrequency(static_cast<uint32>(edgeFreq));
+        exitCopyFallthru->PushBackSuccFreq(static_cast<uint64>(edgeFreq));
         size_t trueBranchIdx = static_cast<uint32>(bb.GetSuccIndex(trueBranch));
-        int64_t updatedtrueFreq = static_cast<int64_t>(bb.GetSuccFreq()[trueBranchIdx] - edgeFreq);
+        int64_t updatedtrueFreq = static_cast<int64_t>(
+            bb.GetSuccFreq()[trueBranchIdx] - static_cast<uint64>(edgeFreq));
         ASSERT(updatedtrueFreq >= 0, "sanity check");
         bb.SetSuccFreq(static_cast<int>(trueBranchIdx), static_cast<uint64>(updatedtrueFreq));
-        pred.AddSuccFreq(edgeFreq, index);
+        pred.AddSuccFreq(static_cast<uint64>(edgeFreq), index);
       }
       return true;
     }
@@ -3993,20 +3994,20 @@ bool ValueRangePropagation::RemoveTheEdgeOfPredBB(
     newBB->AddMeStmtLast(gotoMeStmt);
     PrepareForSSAUpdateWhenPredBBIsRemoved(pred, bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
     size_t index = FindBBInSuccs(pred, bb);
-    uint64 edgeFreq = 0;
+    int64_t edgeFreq = 0;
     if (func.GetCfg()->UpdateCFGFreq()) {
-      edgeFreq = pred.GetSuccFreq()[index];
+      edgeFreq = static_cast<int64_t>(pred.GetSuccFreq()[index]);
       ASSERT(bb.GetFrequency() >= edgeFreq, "sanity check");
     }
     pred.RemoveSucc(bb);
     pred.AddSucc(*newBB, index);
     newBB->AddSucc(trueBranch);
     if (func.GetCfg()->UpdateCFGFreq()) {
-      bb.SetFrequency(bb.GetFrequency() - edgeFreq);
+      bb.SetFrequency(static_cast<uint32>(bb.GetFrequency() - edgeFreq));
       bb.UpdateEdgeFreqs();
-      newBB->SetFrequency(edgeFreq);
-      newBB->PushBackSuccFreq(edgeFreq);
-      pred.AddSuccFreq(edgeFreq, index);
+      newBB->SetFrequency(static_cast<uint32>(edgeFreq));
+      newBB->PushBackSuccFreq(static_cast<uint64>(edgeFreq));
+      pred.AddSuccFreq(static_cast<uint64>(edgeFreq), index);
     }
     DeleteThePhiNodeWhichOnlyHasOneOpnd(bb, updateSSAExceptTheScalarExpr, ssaupdateCandsForCondExpr);
     (void)func.GetOrCreateBBLabel(trueBranch);
